\name{zigguserHeaderCpp-package}
\alias{zigguserHeaderCpp-package}
\alias{zigguserHeaderCpp}
\docType{package}
\title{\packageTitle{zigguserHeaderCpp}}
\description{\packageDescription{zigguserHeaderCpp}}
\section{Package Content}{\packageIndices{zigguserHeaderCpp}}
\author{\packageAuthor{zigguserHeaderCpp}}
\section{Maintainer}{\packageMaintainer{zigguserHeaderCpp}}
\keyword{package}
